requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-restx==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
Werkzeug==2.0.3
